206: Fix UnexpectedDeviceException caused by incorrect Context usage
In most situations the Test and Target context agree on the device metrics.

However, when the user changes properties of the Activity under test at runtime,
such as the orientation of the Activity, the device metrics associated with the
Target Context are changed but the Test Context metrics are not changed.

This means that it is important to use the appropriate context when generating
the device key.